Automatically use a continued MESSAGE_COMMAND if the total command length is larger than 64KB (minus token overhead).  The remctl client library can now send arbitrarily large commands, at some cost in memory consumption on the client and server.  The server is still limited by the OS-imposed maximum length of a command line.
Use the same limit (1MB) on token size everywhere.  Enforce the
protocol limit on unencrypted data size (64KB) in both the server and
when sending messages in the client.

Correctly handle a zero-length argument at the end of a command in the
server.  Previously, that argument was ignored.

Check that the expected argument count matches the count of arguments
seen in the server and that all of the client data was consumed when
parsing arguments.