Fix coding style and API in the PHP PECL module
Clean up the coding style for the PHP PECL module to match the rest of
the project and add copyright headers and license pointers.

Change the extended API exposed by the PECL module to match the API
exposed by other extensions.  This means always providing a data key
with the data, not turning it into stdout and stderr similar to the
simple API, and not adding a done key.  This can be reconsidered later
if there's a very native-PHP way of expressing things, but these
differences seemed unnecessary and not specifically driven by being more
PHP-native.