Improve default arguments and argument parsing for Python bindings
Allow None to be passed in as the principal to the _remctl.remctl
function, and make the port number and principal optional parameters
for _remctl.remctl_open.

Don't set default values for the parameters to remctl(); all parameters
are required.  Likewise for the first element of Remctl.open(), which
is required.  Set a default port of 0 and a default principal of None
so that the library defaults can take effect, rather than hard-coding
defaults.

Allow sending a command with only one value.

Adjust the test suite for new exceptions and error messages.