Python: flesh out setup.py some more
Add classifiers, license information, a long description, and more
metadata keys to setup.py.  Most of this is fairly useless since it's
not being distributed separately, but it can't hurt.  Also provide a
MANIFEST.in so that python setup.py sdist produces the right thing,
even if we'll likely not use that.