Java: improve the Makefile and build.xml file
Rework the build.xml file considerably to more closely follow the
examples in the Ant manual and the recommended file layout for Jakarta.
Change the names of many of the targets for similar reasons.  Include
the examples and configuration files in the distribution created by
ant tar.

Create a manifest file for the remctl JAR that specifies the version
number and origin and identifies the RemctlClient class as the main entry
point.

Use JAVA_HOME in the Makefile instead of JDK since it's more commonly
already set in the environment.

Update the documentation for the new paths used by ant, for the new
capability of generating a distribution, and to follow more closely the
layout used by the other bindings directories.