Further clarify configuration order and deny
Mention more prominantly that the first matching configuration entry will
be used.  Clarify the discussion of ALL and no service paramenter.

Add a caution about using deny with included directories, since the order
in which files in the directory are read is undetermined.

Add sorting the entries in the directory to the to-do list.