Run phpize during configure, update documentation
Don't run phpize from autogen, since the build system that it installs
may vary based on one's local PHP installation.  Instead, run it during
configure.  This means that we don't have to ship a copy of the GPL and
can remove a lot of files from LICENSE.

Update the prerequisites for the PHP bindings and document them for the
Python bindings as well.

Update README for all of the bindings, to point to the protocol
documentation, to mention that the design document is the original design
and is out of date, and to be clearer about prerequisites.