Use command and subcommand for components of a remctl command
Use command and subcommand as the names for the first two parameters
to the remctl client and the first two strings in a remctl command
instead of the unintuitive "type" and "service" terminology borrowed
from sysctl.  This only changes documentation and some internal
variable names; no external APIs should be affected.