Redo server option parsing and logmask checking
Redo parsing of options in the server configuration file to set up the
framework for adding additional options.

Better check logmask options when parsing the server configuration
file and report errors instead of silently ignoring them.  Masking the
command is also no longer supported (it previously worked by
accident).

Improve the documentation of the meaning of the logmask option argument.