Hack the Python test suite to test the right version
Part of the Python build process is to generate multiple build
directories, one for each supported Python version.  The code set
PYTHONPATH when running tests didn't take into account the possibility
of multiple directories.  Modify it to only add the directory for the
current default version of Python.

This will need to be fixed more cleanly upstream.