Python: Add build system hooks for the Debian packaging
Properly building Python modules for Debian requires a few wrinkles, since
Debian needs to pass a non-standard option to setup.py install to install
in the vendor paths and needs to build for multiple versions.  Add a few
hooks to avoid needing to modify Makefile on the fly during packaging.

If PYTHON_INSTALL is set, its contents are passed to the setup.py install
call.

If PYTHON_VERSION is set, the directories to use as a module path during
make check are limited to ones matching that Python version.  This is used
to pick up the right modules for the current default version of Python
when running the test suite.