Coding style and other cleanups for Ruby bindings
Correct the README somewhat on terminology and on which exceptions are
thrown.  Remove (required) notation for arguments when all arguments are
required.  Standardize (hopefully) on correct Ruby notation for methods.

Remove gcc warning options from the CFLAGS in the Ruby extension.  We'll
deal with that another way.

Declare all functions except Init_remctl static and rename Init_remctl
to Init_Remctl to match the name of the class.  Handle a few more places
out of memory errors are possible.