Improve error handling for config file inclusion
Check for opendir errors when including a directory and error out
cleanly instead of segfaulting if the directory disappears between
the stat and our attempt to open it or if we don't have permissions.
Include the system error if stat on an included file fails, since it
might be something interesting like a permission error instead of
just file not found.