When binding IPv6 addresses, use IP_FREEBIND when available
IPv6 addresses can take some time to configure after a system comes
up due to router autoconfiguration, and may not always be available.
Normally, binding to an address that's not yet configured will fail
with an error.  Use IP_FREEBIND where available to bind to addresses
that may not be configured to avoid this problem.

We lose error handling from incorrect addresses, but we gain
robustness against slow network autoconfiguration.