Refactor server token handling
Refactor how server token handling is done to be a bit clearer and
not treat everything else as weird exception cases for command
handling.  Introduces a new error code (protocol spec changes coming
separately) for receiving unexpected tokens in the middle of a
continuation command, and hammers out the continuation handling so
that it's a bit more robust.  (This still needs tests.)