Improve the protocol specification for continued commands
Be clearer about how continued commands are handled.  Forbid messages
other than MESSAGE_QUIT or a valid MESSAGE_COMMAND continuation after
the first part of a continued command.  Be clear that the partial
command must be discarded if anything goes wrong while receiving the
continued command or if a MESSAGE_QUIT was received.

Introduce the ERROR_UNEXPECTED_MESSAGE reply, which can be used for
replying to bad packets sent during a command continuation.