Add additional build system hooks for Debian packaging easier
REMCTL_PERL_FLAGS are additional flags to pass to Makefile.PL when
building the Makefile.

REMCTL_PYTHON_INSTALL adds additional flags to the setup.py install
command.  REMCTL_PYTHON_VERSIONS, if set, is a space-separated list of
Python interpretors with which to build the Python extension.  The
extension will be built, tested, and installed with each one.  If not set,
it will only be built with python (without any version extension).

REMCTL_RUBY controls which Ruby interpretor is used to do the build.

These replace the previous RUBY, PYTHON_INSTALL, and PYTHON_VERSION
variables, which were not as useful.

Change-Id: I9896dcf842482b04a2ad077ce019d332f969a05d