Improve the test infrastructure for spawning remctld
Teach remctld_start to use test_file_path to find the config file,
and have it register remctld_stop as an atexit handler so that it
doesn't have to be called.  Update all callers appropriately.

There are also a few more fixes mixed in to use the new bail memory
allocation functions and to ensure that child helper processes for
tests don't run atexit handlers.

Change-Id: I6bb8096352c70bc81b4101c62e15bd2f61246d84