Map timeout errors to a token error code
Set up a general framework for mapping certain errno values to token
error codes, also used for EOF, and map the timeout error to the
corresponding token error code so that special action can be taken
where appropriate.

Update the logging functions for client and server to recognize the
new token error code.

Change-Id: Ibc2a38574657273ec92c9af5806177ac8f11556a