Add remctl_set_timeout to the client library
Add new remctl_set_timeout function to the remctl library API. Call
this function after remctl_new and before remctl_open to set a network
timeout in seconds.  The client must then receive a reply from the
server in no more than that number of seconds or will abort whatever
action is in progress with a timeout error.

Change-Id: Ica8276460cc551d7fc07d2f6bfcc605d694780da