Add server configuration option to run as a different user
Add a new server configuration option for remctl commands, user,
which sets the user as which to run the command.  If this option is
set, remctld changes to the given user and that user's groups before
running the command.  The groups are set via initgroups and the
primary group is looked up via getpwuid.  The user may be either a UID
or a username.