Add support for running remctld under fakeroot
Add a new function, remctld_start_fakeroot, that starts remctld
under fakeroot.  Support retrieving the actual PID for this case
and the valgrind case so that we can shut down the correct remctld.
Redo the handling of PATH_REMCTLD (and handle PATH_FAKEROOT the
same) so that compilation doesn't fail if it's not defined.

Change-Id: Id7e352ed443f218ac3ef195d5935543cfa58d356