Stop using feature-test macros by default
Do not attempt to use feature-test macros unless compiled with gcc's
-ansi flag or PEDANTIC is defined.  Defining _XOPEN_SOURCE 600 on
Solaris 10, which is required for isinf() if feature-test macros are
in use, causes compilation to fail unless the compiler is run in C99
mode due to an ABI change.  The basic TAP library doesn't care about
this distinction and doesn't want to force specific compiler flags, so
avoiding feature-test macros is a better fix.  Thanks to Bob
Friesenhahn for the bug report.