Fix GSS-API header probing with non-default headers
Fix GSS-API header probes when configure was told to build with a
specific GSS-API library in a non-default path.  Previously, configure
still used the compiler to probe for the correct header names, which
could pick up incorrect headers from the default include path.  Thanks
to Jeffrey Hutzelman for the suggested solution.

Also clean up the way library probes are handled by importing the new
GSS-API probes that will be in rra-c-util 4.3.

Change-Id: Ic94b02ef59004dad191bcd76d9e8eb6e92f4c982
Reviewed-on: https://gerrit.stanford.edu/277
Tested-by: Russ Allbery <rra@stanford.edu>
Reviewed-by: Russ Allbery <rra@stanford.edu>