tests/util/gss-tokens: avoid sequence number mismatch
The test for sending a token with timeout effectively constructs a GSS weap
token and then throws it away without doing anything with it.  This gets the
message token sequence numbers out of sync on opposite ends of the GSS-API
context.  Normally this would not be a problem, but some older versions of
Heimdal require strict sequence number ordering on GSS wrap and MIC tokens,
even if message sequencing and replay detection were not requested.  As a
result, after this test the server side of the context will never be able
to successfully receive a token.

To avoid breaking later tests, the tests for sending and receiving tokens
with timeouts and other problems are moved to the end of the test sequence,
where there is nothing left to break.