Fix various syntax errors and bugs in Perl test suite
The process of making everything comply with perlcritic introduced
a few bugs that needed fixing.  Don't destroy the path for remctld
when spawning one for testing, since we need the full path to find
sleep to run one of the tests.  Take another pass at cleaning up the
code to run klist in stanford-netdb.  (This should all be replaced
by a testing library routine.)  Fix a few other missing imports and
code ordering issues.

Change-Id: Ic64eec456af3033f11506d0f92ac3f16d7bc8dd8
Reviewed-on: https://gerrit.stanford.edu/701
Reviewed-by: Russ Allbery <rra@stanford.edu>
Tested-by: Russ Allbery <rra@stanford.edu>