Add more help output support in Net::Remctl::Backend
Derive the command name from the command structure, so it doesn't
have to be stated in the syntax.  Add new parameters to the
Net::Remctl::Backend constructor to set a banner for the help output
and to set the top-level command so that it can be added to the
syntax.  Tighten up the space given to command summaries.

Change-Id: Ie73f7f64381954028d77165cbe0000554b840d5f
Reviewed-on: https://gerrit.stanford.edu/708
Reviewed-by: Russ Allbery <rra@stanford.edu>
Tested-by: Russ Allbery <rra@stanford.edu>