Move connection-opening to remctl_open()
Take code that selects a port and calls internal_connect() to establish
a connection, and pull it up out of internal_open() and into remctl_open().
In the process, internal_connect() is made non-static, and internal_open()
loses its port argument.

This paves the way for alternatives to remctl_open() in which the caller
provides a socket descriptor or address instead of a hostname and port.
Such interfaces will need internal_open() without internal_connect().

Change-Id: I81d90097a3d422fb9c6a3fe754a5a1f60aac4415
Reviewed-on: https://gerrit.stanford.edu/806
Reviewed-by: Russ Allbery <rra@stanford.edu>
Tested-by: Russ Allbery <rra@stanford.edu>