Switch Perl build system to Module::Build
The Net::Remctl Perl bindings now require Perl 5.8 or later (instead
of 5.006 in previous versions) and are now built with Module::Build
instead of ExtUtils::MakeMaker.  This should be transparent to anyone
not working with the source code, since Perl 5.8 was released in 2002,
but Module::Build and ExtUtils::CBuilder are now required to build
Net::Remctl.  They are included in Perl 5.10 or later and can be
installed separately for older versions of Perl.

Change-Id: Ifc54e38d7efcc01cab5f0ae3c89ebed8e748166a
Reviewed-on: https://gerrit.stanford.edu/771
Reviewed-by: Russ Allbery <rra@stanford.edu>
Tested-by: Russ Allbery <rra@stanford.edu>