Make Perl test suite self-contained
Copy the key bits from the main test suite into the Perl test suite,
either in the distribution or, for test configuration, dynamically on
make check.  Eliminate the need for Autoconf substitution in the API
check by passing in the path to the remctld as an environment variable.
This means that it's now meaningful to create a standalone Perl
distribution of Net::Remctl.

Change-Id: I71def2b71e9da832e7af042c899ebb37ab18df5d
Reviewed-on: https://gerrit.stanford.edu/781
Reviewed-by: Russ Allbery <rra@stanford.edu>
Tested-by: Russ Allbery <rra@stanford.edu>