Make Ruby extension builds verbose
We want the verbose logging for automated builds, such as Debian
package builds (which do compiler and linker flag analysis).  We
could do something more complex based on the setting of verbosity
for the main package, but that's not trivial and it's not very
important to suppress those two lines.  PHP and Perl are much more
verbose anyway.

Change-Id: I15d4590ea03a93f017579a47769492c092e0a290
Reviewed-on: https://gerrit.stanford.edu/1200
Reviewed-by: Russ Allbery <rra@stanford.edu>
Tested-by: Russ Allbery <rra@stanford.edu>