Fix race condition in remctld with output before process end
The backend being run is a shell script that does the following:

        cat /root/tmp/databases/tw.db
        rm -rf /root/tmp/databases
        exit 0

remctld doesn't wait for the backend program to close its standard output
and standard error since, in some cases (such as when the backend program
spawns a daemon via an init script), standard output and standard error
are never closed.  Instead, it waits until the child process exits, and
then considers the command finished.

The problem is that the operating system buffers pipes.  In the above, cat
exits as soon as it's successfully written all of the file to its output
pipe.  However, that success from the operating system just indicates that
all of the output has been buffered.  With newer versions of Linux, that
buffer can be quite large.  It then goes on to run rm and exit.

In the remctld process, on each trip through the select loop, it polls to
see if the child process has exited.  If so, it attempts one more time to
read output data from the process with a timeout of 0, and then considers
the command complete.

The problem is that the remctld server process reads data in chunks of
64000 bytes.  If the file is larger, say 100KB, a modern version of Linux
will buffer the entire file contents in the pipe buffer and cat will
essentially exit as soon as the file contents have been copied into kernel
space (which is extremely fast since it's just a move of pages from the
disk cache to the pipe cache).  It's then possible, if remctld loses the
race, for the rm and exit to finish and for remctld to be signaled before
remctld reads any of the data.  remctld will then observe the child
process has exited, read one more block of 64000 bytes, and then leave the
rest of the data unread in the pipe buffer, resulting in truncated
output.  Usually only a small amount of data in excess of a multiple of
64000 at the end of the file is truncated.

The fix is to reduce the timeout to 0 but continue looping on select with
a timeout of 0 (indicating a poll of the file descriptors) until no
further output is immediately available for reading, and only then
consider the command complete.

Change-Id: I38f77a8a9b9d9938ef0521a3df73ea75399cd0bf
Reviewed-on: https://gerrit.stanford.edu/1207
Reviewed-by: Russ Allbery <rra@stanford.edu>
Tested-by: Russ Allbery <rra@stanford.edu>