Stop substituting Perl module versions in configure
The Perl modules now have static version numbers, which are updated
using the --update option of the tests/perl/module-version.t script
(coming in the next commit).  Remove the configure machinery.

Also make the Perl module rebuild depend on Net::Remctl::Backend.

The version numbers of the Net::Remctl and Net::Remctl::Backend Perl
modules now match the versions of the remctl package, but with at
least two digits for the minor version so that, for example, 3.9
(which becomes 3.09) and 3.10 will sort properly as numbers.  This
means that, from Perl's perspective, the version numbers have gone
backwards in this release relative to earlier 3.0 releases.  This is a
one-time adjustment to a more reliable versioning scheme.

Change-Id: I0a0811f13ca35428cbd97ad9d9fc34f79bb7c56b
Reviewed-on: https://gerrit.stanford.edu/1220
Reviewed-by: Russ Allbery <rra@stanford.edu>
Tested-by: Russ Allbery <rra@stanford.edu>