Don't report the legacy port on connection errors
When we support port fallback (when the port argument to remctl_open
is 0), we first try the current port and then the legacy port.  This
meant that the reported connection failure error message always
specified the legacy port.

Change this by remembering the error from the first connection and
reporting it if the legacy connection also fails.  This avoids
encouraging people to set up firewall rules for port 4444 or otherwise
chasing the wrong problem or using the legacy port where they shouldn't.

Change-Id: I9a5078cbfb61cfeaf9aa017782985b3f6520dd12
Reviewed-on: https://gerrit.stanford.edu/1295
Reviewed-by: Russ Allbery <rra@stanford.edu>
Tested-by: Russ Allbery <rra@stanford.edu>