Fix various distribution, build, and clean issues
The current Ruby build system refuses to generate a proper
Makefile when the source isn't in the same directory, so go ahead
and copy everything over like we do for other language bindings.

Add some missing Perl directories that have to be created for
out-of-tree builds to work properly.

Add the systemd unit files and sd-daemon.h portability header file
to the appropriate variables in Makefile.am.

Pass the correct compiler flags to the Ruby build when building
out of tree so that it can find all the header files.

Change-Id: I27bd87ba3bc42cf3f9ed1b6a5f6d3f918a10fbe6
Reviewed-on: https://gerrit.stanford.edu/1359
Reviewed-by: Russ Allbery <rra@stanford.edu>
Tested-by: Russ Allbery <rra@stanford.edu>