Switch the process event loop to libevent
Replace the complex hand-coded client event loop, using select,
with a rewritten one using libevent.  This fixes race conditions
in noticing the client exit properly and removes the need for the
five-second polling to see whether the client has exited.  It also
provides a framework for eventual added complexity in the event
loop.

Import the buffer utility functions from rra-c-util and use that
to manage the client output buffer for now.  This will probably be
replaced with evbuffers.

This is the first cut, using basic libevent functionality.  The
next step will be to use bufferevents in some places and extend the
use of libevent to other event loops in the server.

Currently requires libevent 2.x.  Portability to libevent 1.4.3
will come in a subsequent commit.

Change-Id: Ia8e4cd0a857c710dbc0c348578648084466e886b
Reviewed-on: https://gerrit.stanford.edu/1378
Reviewed-by: Russ Allbery <rra@stanford.edu>
Tested-by: Russ Allbery <rra@stanford.edu>