Simplify process I/O sockets
Since we're now using socketpair, we can use a single socket for
both standard input and standard output (and, for protocol version
one, standard error as well).  Do this along with the resulting
code simplification and unification of some callbacks.  Handle
ECONNRESET errors alongside EPIPE now that we're doing both input
and output on the same socket.

Change-Id: I960f39e71ca8db9405a0b41b01a334b5e82ed8e9
Reviewed-on: https://gerrit.stanford.edu/1388
Reviewed-by: Russ Allbery <rra@stanford.edu>
Tested-by: Russ Allbery <rra@stanford.edu>