Use an evbuffer to store input data for a process
Start isolating the process event loop from the rest of the code
by not reusing the struct iovec representation of the input data.
Instead, convert it to an evbuffer first.

This does make another copy of the data.  We will rethink this if
it shows up as a performance issue.

Change-Id: Ie37dea0e9b4a58fa4f481ed0a4fd5ba8bb7ebdab
Reviewed-on: https://gerrit.stanford.edu/1389
Reviewed-by: Russ Allbery <rra@stanford.edu>
Tested-by: Russ Allbery <rra@stanford.edu>