Move more child processing into the event loop
The previous implementation of the child process handling raced
the child exit with the event loop setup.  Solve this problem by
setting up the event loop and the SIGCHLD handler first, and then
spawn the process via a one-time event inside the event loop.
This ensures that libevent is ready to catch the SIGCHLD signal
before it could possibly be sent.

Move all of the process handling code into a separate file and
use that refactoring to try to simplify the code and make it more
straightforward.  There's still quite a bit of work to do on this.

Fix an evbuffer leak when processing standard input.

Change-Id: I17c03b9e7ac4fabadd1c8f95d647407a7d0fae18
Reviewed-on: https://gerrit.stanford.edu/1391
Reviewed-by: Russ Allbery <rra@stanford.edu>
Tested-by: Russ Allbery <rra@stanford.edu>