Multiple improvements to test process handling
When spawning subprocesses during a test, clean up the allocated
memory in child processes as well as the primary process (but only
stop the child when called from the primary test process).  Rather
than using alarm to die if the child doesn't stop, wait for it to
exit in a select delay loop, and bail if it doesn't stop.

Change-Id: Ie73f914f2fc1c1082e798630130876f048dcf76f
Reviewed-on: https://gerrit.stanford.edu/1404
Reviewed-by: Russ Allbery <rra@stanford.edu>
Tested-by: Russ Allbery <rra@stanford.edu>