Free all resources on remctld exit or fork
Be more careful about resources and free everything on exit,
particularly the GSS-API credentials, which will close all the
replay cache file handles.  Also free as much as possible when
transferring control to the child process to handle an incoming
connection, and free the rest when that client has been handled.

This allows clean valgrind analysis of the remctld server,
including file descriptor tracking.

Change-Id: Ie4806e5b9374a9e6ad6d3c6c29f4b22fa2247dbc
Reviewed-on: https://gerrit.stanford.edu/1407
Reviewed-by: Russ Allbery <rra@stanford.edu>
Tested-by: Russ Allbery <rra@stanford.edu>