Backport to libevent 1.4
Now builds and works correctly with the version of libevent that
shipped with Debian squeeze.  Add lots and lots of portability
code for all the changed or missing APIs in libevent 1.4.  Stop
relying on event_base_got_break and set a flag in the process
instead, since that function isn't available in 1.4.  Avoid clearing
the event callback on the inout buffevent when switching to discard
processing since this causes a segfault in libevent 1.4.

Change-Id: If75d455667327ee4e509b7fefb989ec7bf1d2dcf
Reviewed-on: https://gerrit.stanford.edu/1413
Reviewed-by: Russ Allbery <rra@stanford.edu>
Tested-by: Russ Allbery <rra@stanford.edu>