Refactoring of the localgroup support
Break the long ACL verification function into three functions
to try to make them a bit easier to read.  Use a larger but simpler
buffer for storing the local username.  Add a check to see if the
given group is the user's primary group rather than a supplemental
group.

Also do some refactoring of the test suite, renaming the fake
getgrnam support to a separate file name and adding fake getpwnam
information as well.

Pull in the messages-krb5 utility library from rra-c-util to make
it easier to report Kerberos errors with parsing the principal or
determining the local username.

Change-Id: I31645f90843cae0e96c88b451beea71790869313
Reviewed-on: https://gerrit.stanford.edu/1503
Reviewed-by: Russ Allbery <rra@stanford.edu>
Tested-by: Russ Allbery <rra@stanford.edu>