Only add warning suppression flags with make warnings
Previous versions always passed the flags to disable certain warnings
to the language binding builds, even if warnings weren't otherwise
enabled.  As of remctl 3.9, that included a warning flag not supported
by old versions of gcc, breaking builds on RHEL 5.  Instead, only pass
the warning suppression flags when building with warnings (via make
warnings), which is not the default and is only supported with recent
versions of gcc.  Thanks to Ken Dreyer for the report.

This also allows removal of the GCC Automake conditional.

Change-Id: I50fa0242931802cad954ec5c5a0d6ab53b8c47e8
Reviewed-on: https://gerrit.stanford.edu/1534
Reviewed-by: Russ Allbery <rra@stanford.edu>
Tested-by: Russ Allbery <rra@stanford.edu>