Skip anonymous tests if Kerberos is too old
Some older versions of Heimdal not only can't do PKINIT, but also crash
when krb5_get_init_creds_password() is called with no password or prompter.
Heimdal versions which avoid the crash have krb5_init_creds_set_password(),
as do all versions of MIT Kerberos which support PKINIT.  So, disable the
anonymous tests if that function is not present.

Note that there is a 4-month window in which the new function is present
in Heimdal and the crash is still possible, but there were no releases
during that time.