Initial refactoring for ssh shell support
The first step towards adding a new server mode that can be run as
a shell via ssh.  Refactor the code to more cleanly separate the
protocol implementation and the GSS-API and non-GSS-API bits, and
add the remctl-shell binary.  This is currently entirely untested,
apart from ensuring that it doesn't break the existing server
implementation.