Add new sudo configuration option
Add a new configuration option, sudo, which tells remctld and
remctl-shell to run the command as a different user using sudo.  The
path to the sudo binary is determined when remctld is compiled.
Normally, it's more convenient to use the existing user option, but it
relies on remctld running as root.  If running the daemon as a
non-root user, or when running remctl-shell as a non-root user, this
option may work better.