Java: fix name canonicalization & GSS name type
Fix a very common mistake: the Java remctl client uses a Kerberos
principal instead of a GSS hostbased-service name, which is generally not
at all what you want since this ignores the local policies on host/realm
mapping and name canonicalization -- and it tries to make up for some of
the resulting problems by canonicalizing the hostname itself. Instead:

  /* Determine the server identity. If the user passed a
   * servicePrincipal, use it verbatim and leave the name type null
   * to select the system default interpretation (which will
   * generally by a Kerberos principal name). If the user did not
   * specify, our default is the hostbased-service name
   * "host@<hostname>", which Kerberos will map to a principal name
   * using the configured policy regarding name/realm mapping,
   * canonicalization, and so forth. Note that we do *NOT*
   * canonicalize or otherwise transform the hostname. Applications
   * should pass the user's intent unchanged to GSSAPI and let the
   * system configuration do what it's intended to do -- e.g. in MIT
   * Kerberos the krb5.conf 'rnds' parameter controls whether
   * reverse-address mapping is done on hostnames; we should not
   * override that.
   */

This tweak doesn't add a way for the user to pass a custom
hostbased-service name; usually if you want to override, you know what
principal you want.

(This is all what the C client does.)