Update to rra-c-util 7.1 and C TAP Harness 4.3
Validate command argument count, the length of command arguments, and
the length of blocks of output from the server fit into the data type
used in the wire protocol.

Check the port argument to remctl and remctld to ensure that it is a
valid port number.

Update to rra-c-util 7.1:

* Avoid spurious test failures from the network library.
* Fix configure output when a Kerberos install prefix was provided.
* Fix new warnings in GCC 7 and add new warning flags.
* Fix all warnings from the Clang static analyzer.
* Fix warnings under Clang with most warnings enabled.

Update to C TAP Harness 4.3:

* On test failures, report left/right instead of wanted/expected.
* Fix string comparisons involving NULL pointers.