Convert documentation to DocKnot
Move the THANKS section to a separate file to try to reduce the
length of the top-level README a bit.  Start providing a Markdown
README.md as well for GitHub.  Tweak a lot of documentation wording
and move things around to bring the package documentation in line
with my standard templates.