Remove obsolete Stanford-only test
The remctl Perl API had a test for explicit undef that required
remctld be running on the standard port, and thus tested against
a Stanford-internal service (but only if one had stanford.edu
Kerberos credentials).  Essentially no one runs this test any
longer, and the testing value is not that large, so drop it rather
than keeping it up to date.

The correct way to do this test would be with a more complex
testing setup involving containers that could start remctld on the
standard port, but it's not worth it to add such a test environment
only for this test.