Fix precision of snprintf floating point tests
32-bit floating point numbers only have 7 digits of precision, but
the test suite was testing with 9 digits.  Clang produces different
results in the Travis-CI environment from system snprintf in the
eighth and ninth digits, which seems to be allowed by the floating
point precision.  Since the snprintf replacement is rarely used,
just lower the precision rather than try to debug it further.