Streamline the CI build
Rather than adding the compiler to the matrix, build first with
Clang and then with GCC and only use Kerberos in the matrix.  This
will make each build take a bit longer, but consume about half as
many resources.

Add cppcheck to the build for the MIT Kerberos variant (no point in
running it under both types of Kerberos).