Overhaul GitHub Actions testing
Use local::lib and cache Perl modules to speed up installation
time when repeatedly testing.  Disable testing of the Perl modules,
which should speed up installation considerably.  Move the list of
apt packages to install to a separate file for easier maintenance.

Fix the list of test dependencies, which was missing some of the
modules that the Perl binding test suite does use.