Update extension warning suppression
Update the warning suppression rules for all of the language
extensions for changes to the compilers. Ruby required some
special attention because Ruby defines a possibly-conflicting
`-D_FORTIFY_SOURCE` flag and GCC errors if the same symbol is
defined twice on the command line. Add a bit of shell code to
`Makefile.am` to strip out that flag when building the Ruby
extension.

Change `_FORTIFY_SOURCE` to 3 in the GCC warning flags.